# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your Super Application API",
        "description": "API Documentation for the Booking Courses application",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "www.apache.org"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get all courses",
                "operationId": "a111247eb122c574b1bdf2bfec710c8a",
                "responses": {
                    "200": {
                        "description": "Courses retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Laravel Basics"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Learn Laravel from scratch"
                                            },
                                            "level": {
                                                "type": "string",
                                                "example": "Beginner"
                                            },
                                            "total_seats": {
                                                "type": "integer",
                                                "example": 20
                                            },
                                            "available_seats": {
                                                "type": "integer",
                                                "example": 20
                                            },
                                            "rating": {
                                                "type": "number",
                                                "example": 4.5
                                            },
                                            "duration": {
                                                "type": "string",
                                                "example": "3 hours"
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "image_url": {
                                                "type": "string",
                                                "example": "courses/image.jpg"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Request failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Create a new course",
                "operationId": "b0442045a77247128649ebcb0c3526a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "level",
                                    "total_seats",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Laravel Basics"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Learn Laravel from scratch"
                                    },
                                    "level": {
                                        "type": "string",
                                        "example": "Beginner"
                                    },
                                    "total_seats": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "available_seats": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "rating": {
                                        "type": "number",
                                        "example": 4.5
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "3 hours"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image_url": {
                                        "description": "Course image file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Laravel Basics"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Learn Laravel from scratch"
                                        },
                                        "level": {
                                            "type": "string",
                                            "example": "Beginner"
                                        },
                                        "total_seats": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "available_seats": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "rating": {
                                            "type": "number",
                                            "example": 4.5
                                        },
                                        "duration": {
                                            "type": "string",
                                            "example": "3 hours"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "image_url": {
                                            "type": "string",
                                            "example": "courses/image.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Request failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}